+++ dependencies/samcodes-notifications/src/com/samcodes/notifications/BootReceiver.java
@@ -39,35 +39,6 @@ public class BootReceiver extends BroadcastReceiver {
 	}
 	
 	private static void registerAlarmsPostBoot(Context context) {
-		Common.pendingIntents.clear(); // Clear out the pending intents, precaution in case spurious/multiple post-boot events are received somehow
-		
-		Long currentTime = System.currentTimeMillis();
-		
-		// Alarm re-registration. This is required because alarms are cleared when the device is turned off or rebooted
-		Log.i(Common.TAG, "Re-registering application notifications on boot");
-		for(int slot = 0; slot < Common.MAX_NOTIFICATION_SLOTS; slot++) {
-			SharedPreferences prefs = Common.getNotificationSettings(context, slot);
-			if(prefs == null) {
-				continue;
-			}
-			Long alertTime = prefs.getLong(Common.UTC_SCHEDULED_TIME, -1);
-			if(alertTime == -1) {
-				continue; // Skip unreadable/not-set notification data
-			}
-			if(alertTime - currentTime < 0) {
-				// Reschedule notifications whose time passed while the phone was powered off to the very-near future, preserving order
-				double overdueByMillis = Math.abs(alertTime - currentTime);
-				double orderPreservingDelay = 100 + (1000 / (1 + Math.log10(overdueByMillis + 1)));
-				alertTime = currentTime + Double.valueOf(orderPreservingDelay).longValue();
-			}
-			
-			String titleText = prefs.getString(Common.TITLE_TEXT_TAG, "");
-			String subtitleText = prefs.getString(Common.SUBTITLE_TEXT_TAG, "");
-			String messageBodyText = prefs.getString(Common.MESSAGE_BODY_TEXT_TAG, "");
-			String tickerText = prefs.getString(Common.TICKER_TEXT_TAG, "");
-			
-			PendingIntent intent = Common.scheduleLocalNotification(context, slot, alertTime, titleText, subtitleText, messageBodyText, tickerText);
-			Common.pendingIntents.put(slot, intent);
-		}
+		Common.reregisterAlarms(context);
 	}
 }